requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.2
imageio==2.27.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
networkx==3.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tifffile==2023.3.21
zipp==3.15.0

torch~=2.0.0
torchvision~=0.15.1
torchmetrics~=0.11.4
PyYAML~=6.0
kornia~=0.6.11
open3d~=0.17.0
tqdm~=4.65.0
joblib~=1.2.0